improving patient dashboard

+++ patients/views.py
@@ -3,7 +3,7 @@ from .models import Patient
 from .forms import PatientForm 
 from accounts.forms import RegisterForm
 from django.contrib.auth.decorators import login_required
-
+from django.contrib.auth import login
 
 def patients_list(request):
     patients = Patient.objects.all()
@@ -22,6 +22,7 @@ def patient_form(request):
             patient_profile = patient_form.save(commit=False)
             patient_profile.user = user
             patient_profile.save()
+            login(request, user)
             return redirect("patient_dashboard")   
     else: 
         patient_form = PatientForm()
@@ -31,10 +32,14 @@ def patient_form(request):
 
 def patient_detail(request,pk):
     patient = get_object_or_404(Patient, pk=pk)
+    if patient.user != request.user:
+        return redirect("patients_list")
     return render(request,"patients/patient_detail.html",{ "patient": patient })
 
 def patient_modify(request,pk):
     patient= get_object_or_404(Patient, pk=pk)
+    if patient.user != request.user:
+        return redirect("patients_list")
     if request.method == "POST":
         patient_form = PatientForm(request.POST,instance=patient)
         user_form = RegisterForm(request.POST,instance=request.user)
@@ -50,10 +55,18 @@ def patient_modify(request,pk):
 @login_required
 def patient_delete(request,pk):
     patient= get_object_or_404(Patient, pk=pk)
+    if patient.user != request.user:
+        return redirect("patients_list")
     if request.method == "POST":
         patient.delete()
         return redirect("patients_list")
     return render(request,"patients/patient_confirmation.html",{"patient":patient})
 
 def patient_dashboard(request):
-    pass
+    patient = request.user.patient_profile
+    appointments = patient.appointments.all()
+    total_appointments = appointments.count()
+    if patient.user != request.user:
+        return redirect("patients_list")
+    context = {"patient":patient, "appointments":appointments, "total_appointments":total_appointments}
+    return render(request,"patients/patient_dashboard.html",context)
